Used (size_t), instead of (long) where converting between pointers and integers.
(long) still is 32 bits on 64-bit Windows!

+++ src/cc65/symentry.h
@@ -115,7 +115,7 @@ struct CodeEntry;
 typedef struct DefOrRef DefOrRef;
 struct DefOrRef {
     unsigned            Line;
-    long                LocalsBlockId;
+    size_t              LocalsBlockId;
     unsigned            Flags;
     int                 StackPtr;
     unsigned            Depth;
